checkeo en python ahora usa int8

+++ cuda/sorter.py
@@ -1,7 +1,7 @@
 import numpy as np
 
-sin_ordenar = np.fromfile('sin_ordenar', dtype=np.int32)
-ordenado = np.fromfile('ordenada', dtype=np.int32)
+sin_ordenar = np.fromfile('sin_ordenar', dtype=np.int8)
+ordenado = np.fromfile('ordenada', dtype=np.int8)
 
 ordenado_python = np.sort(sin_ordenar)
